Introduce 'accounts' app for user-specific views & resources
