+++ netbox/templates/account/preferences.html
@@ -1,4 +1,4 @@
-{% extends 'users/account/base.html' %}
+{% extends 'account/base.html' %}
 {% load helpers %}
 {% load form_helpers %}
 {% load i18n %}
