+++ netbox/templates/account/profile.html
@@ -1,4 +1,4 @@
-{% extends 'users/account/base.html' %}
+{% extends 'account/base.html' %}
 {% load helpers %}
 {% load render_table from django_tables2 %}
 {% load i18n %}
