+++ netbox/templates/account/token.html
