--- netbox/templates/users/account/password.html
@@ -1,21 +1,0 @@
-{% extends 'users/account/base.html' %}
-{% load form_helpers %}
-{% load i18n %}
-
-{% block title %}{% trans "Change Password" %}{% endblock %}
-
-{% block content %}
-    <form action="." method="post" class="form form-horizontal col-md-8 offset-md-2">
-        {% csrf_token %}
-        <div class="field-group">
-            <h5 class="text-center">{% trans "Password" %}</h5>
-            {% render_field form.old_password %}
-            {% render_field form.new_password1 %}
-            {% render_field form.new_password2 %}
-        </div>
-        <div class="text-end">
-            <a href="{% url 'account:profile' %}" class="btn btn-outline-danger">{% trans "Cancel" %}</a>
-            <button type="submit" name="_update" class="btn btn-primary">{% trans "Save" %}</button>
-        </div>
-    </form>
-{% endblock %}
