--- netbox/templates/users/account/token_list.html
@@ -1,26 +1,0 @@
-{% extends 'users/account/base.html' %}
-{% load helpers %}
-{% load render_table from django_tables2 %}
-{% load i18n %}
-
-{% block title %}{% trans "My API Tokens" %}{% endblock %}
-
-{% block content %}
-<div class="row">
-	<div class="col col-md-12 text-end">
-    <a href="{% url 'account:usertoken_add' %}" class="btn btn-sm btn-primary my-3">
-      <span class="mdi mdi-plus-thick" aria-hidden="true"></span> {% trans "Add a Token" %}
-    </a>
-  </div>
-</div>
-<div class="row mb-3">
-	<div class="col col-md-12">
-    <div class="card">
-      <div class="card-body table-responsive">
-        {% render_table table 'inc/table.html' %}
-        {% include 'inc/paginator.html' with paginator=table.paginator page=table.page %}
-      </div>
-    </div>
-  </div>
-</div>
-{% endblock %}
